requirements.txt
appdirs==1.4.4
audioread==2.1.9
backcall==0.2.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
decorator==5.0.9
docutils==0.17.1
ffmpeg==1.4
ffprobe==0.5
idna==2.10
ImageHash==4.2.0
importlib-metadata==4.3.0
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
keyring==23.0.1
kiwisolver==1.3.1
librosa==0.8.1
llvmlite==0.36.0
matplot==0.1.9
matplotlib==3.4.2
matplotlib-inline==0.1.2
numba==0.53.1
numpy==1.20.3
packaging==20.9
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
pkginfo==1.7.0
pooch==1.3.0
prompt-toolkit==3.0.18
pycparser==2.20
pydub==0.25.1
Pygments==2.9.0
pyloco==0.0.139
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.1
PyWavelets==1.1.1
pywin32-ctypes==0.2.0
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
resampy==0.2.2
rfc3986==1.5.0
scikit-learn==0.24.2
scipy==1.6.3
SimpleWebSocketServer==0.1.1
six==1.16.0
SoundFile==0.10.3.post1
threadpoolctl==2.1.0
tqdm==4.61.0
traitlets==5.0.5
twine==3.4.1
typing==3.7.4.3
urllib3==1.26.5
ushlex==0.99.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.0.1
zipp==3.4.1
